backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pytesseract
Pillow
vaderSentiment
